use class initializer for (s)ftp server
